refactor: adjust configuration and application starts
