+++ packages/utils/errors.ts
@@ -1,4 +1,4 @@
-import config from 'config';
+import config from '../../config';
 import Logger from './logging';
 
 const logger = Logger('errors');
