+++ server.ts
@@ -1,11 +1,11 @@
-import config from 'config';
+import config from './config';
 import { createExpressApplication } from './app';
 import gracefulShutdown from '@nc/utils/graceful-shutdown';
 import Logger from '@nc/utils/logging';
 import { createServer as createHTTPServer, Server } from 'http';
 import { createServer as createHTTPSServer, Server as SecureServer } from 'https';
 
-export function createServer() {
+(function() {
   const logger = Logger('server');
   const app = createExpressApplication();
   const server: Server | SecureServer = (config.https.enabled === true) ? createHTTPSServer(config.https, app as any) : createHTTPServer(app as any);
@@ -16,4 +16,4 @@ export function createServer() {
   });
 
   return server;
-}
+}());
