+++ test/utils/package.json
@@ -6,7 +6,6 @@
   "devDependencies": {
     "@nc/utils": "0.0.0",
     "@types/jest": "^26.0.0",
-    "config": "^3.3.0",
     "jest": "^27.0.0",
     "ts-jest": "^27.0.0"
   }
